♻️ refactor(tarn): split interpreter.rs into errors/frame/calls/selectors, pool call-arg Vecs
interpreter.rs had grown into a 2578-line monolith mixing VmError, frame/pool
state, call/parameter-binding machinery, selector evaluation, and the opcode
dispatch loop. Split the first four into their own submodules (mirroring the
front-door pattern tarn.rs already uses for cache/nodes_split/disasm), moving
each concern's tests alongside it. Also unified eval_selector_expr and
eval_selector_expr_with_args, which had drifted into near-duplicate Array/Dict
recursion, into one implementation parameterized on an optional args slice.

Separately, call_stack_value's non-fixed-arity path (variadic/optional params,
partial()-bound closures) now pulls its argument Vec from the existing
operand-stack pool instead of allocating fresh on every call.

Verified across default/tarn/tarn+debugger/tarn+debugger+debug-trace/tarn+sync/
--all-features: cargo check, clippy -D warnings, fmt --check, nextest (5449
tests), and the bundled builtin_tests.mq suite run through Tarn all pass.
Benchmarks show no regression on the hot paths (fibonacci, foreach, array
map/filter/fold).
